refactor: move glob util to internal/utils
The glob matcher lived under the `solidity-test` plugin's `eip712`
folder, but it has no eip712 specifics — it is a general path-matching
utility. We move it to `internal/utils/glob.ts` so other built-in
plugins can reuse it, and expose a `matchesAnyGlob` helper for the
plain "does this path match any of these globs" case that
`isPathSelected`'s include/exclude gating does not fit.

This is a pure move with no behaviour change.
